split json and jsonp tests and add test cases with checking still supported syntax
